- rename tests/00-runtime/runtime.js 	- tweak aforementioned test file 	- add readFile() to lib/stdio re: issue #71 	- add tests for fopen, write, readFile()
